requirements.txt
# General dependencies
pyyaml==6.0.2

# LLM dependencies
openai==1.65.1

# API
flask==3.1.0
requests==2.32.3
flask-cors==5.0.1